moved is_primitive/is_structured functions into archytypes. structured types is generally working like before
